Pull out hoist code into modules
